updated tests to check cases of incorrect input, moved kafkaoption to cli_utils
